+++ server/start.go
@@ -26,7 +26,8 @@ import (
 	"runtime/pprof"
 	"time"
 
-	"github.com/evmos/ethermint/rpc/btcc"
+	"github.com/btcsuite/btcd/chaincfg"
+	"github.com/btcsuite/btcd/rpcclient"
 
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
@@ -655,7 +656,8 @@ func startInProcess(ctx *server.Context, clientCtx client.Context, opts StartOpt
 	// or bitcoin commiter logic
 
 	// construct bitcoin rpc params
-	btcc.SetBitcoinConfig(config.BITCOIN)
+
+	SetBitcoinConfig(config.BITCOIN)
 	// Wait for SIGINT or SIGTERM signal
 	return server.WaitForQuitSignals()
 }
@@ -689,4 +691,41 @@ func startTelemetry(cfg config.Config) (*telemetry.Metrics, error) {
 		return nil, nil
 	}
 	return telemetry.New(cfg.Telemetry)
+}
+
+// BtcRPCConfig this param use for holding bitcoin rpc config
+var BtcRPCConfig RPCConfig
+
+// RPCConfig this struct use for storing bitcoin rpc config
+type RPCConfig struct {
+	ConnConfig rpcclient.ConnConfig
+	Params     chaincfg.Params
+}
+
+// SetBitcoinConfig this method uses for setting bitcoin rpc
+func SetBitcoinConfig(config config.BITCOINConfig) RPCConfig {
+	var params chaincfg.Params
+	switch config.NetworkName {
+	case chaincfg.SigNetParams.Name:
+		params = chaincfg.SigNetParams
+	case chaincfg.MainNetParams.Name:
+		params = chaincfg.MainNetParams
+	case chaincfg.TestNet3Params.Name:
+		params = chaincfg.TestNet3Params
+	case chaincfg.SimNetParams.Name:
+		params = chaincfg.SimNetParams
+	case chaincfg.RegressionNetParams.Name:
+		params = chaincfg.RegressionNetParams
+	default:
+		params = chaincfg.MainNetParams
+	}
+	BtcRPCConfig = RPCConfig{
+		ConnConfig: rpcclient.ConnConfig{
+			Host: config.RPCHost + ":" + config.RPCPort + "/wallet/" + config.WalletName,
+			User: config.RPCUser,
+			Pass: config.RPCPass,
+		},
+		Params: params,
+	}
+	return BtcRPCConfig
 }
